+++ src/main/java/testing/shoppingCart/CashPayment.java
@@ -4,7 +4,7 @@ import java.util.Map;
 
 public class CashPayment extends Billing implements Payment,Discount {
     @Override
-    public void paymentCalculation(Map<Product, Integer> listOfProducts) {
+    public void paymentCalculation(Map<Product, Double> listOfProducts) {
         generateBill(listOfProducts);
         System.out.println("5 % discount on wallet payment on payment over 100, Total: "
                 +(total-total*discountOffer())+"\n");
