+++ src/main/java/testing/shoppingCart/Payment.java
@@ -3,5 +3,5 @@ package testing.shoppingCart;
 import java.util.Map;
 
 public interface Payment {
-    void paymentCalculation(Map<Product, Integer> listOfProducts);
+    void paymentCalculation(Map<Product, Double> listOfProducts);
 }
